+++ src/python/siarray.rs
@@ -1,7 +1,7 @@
 use crate::si::*;
 use crate::QuantityError;
 use numpy::{PyReadonlyArray1, PyReadonlyArray2, PyReadonlyArray3, PyReadonlyArray4, ToPyArray};
-use pyo3::exceptions::PyTypeError;
+use pyo3::exceptions::{PyIndexError, PyTypeError};
 use pyo3::prelude::*;
 use pyo3::{PyNumberProtocol, PySequenceProtocol};
 use std::ops::Deref;
@@ -73,5 +73,49 @@ impl PySIArray1 {
     #[pyo3(text_signature = "(start, end, n)")]
     fn logspace(start: PySINumber, end: PySINumber, n: usize) -> Result<Self, QuantityError> {
         Ok(SIArray1::logspace(start.0, end.0, n)?.into())
+    }
+}
+
+#[pyproto]
+impl PySequenceProtocol for PySIArray1 {
+    fn __len__(&self) -> PyResult<usize> {
+        Ok(self.0.len())
+    }
+
+    fn __getitem__(&self, idx: isize) -> PyResult<PySINumber> {
+        if idx < 0 || idx >= self.0.len() as isize {
+            Err(PyIndexError::new_err("array index out of bounds"))
+        } else {
+            Ok(PySINumber(self.0.get(idx as usize)))
+        }
+    }
+
+    fn __setitem__(&mut self, idx: isize, value: PySINumber) -> PyResult<()> {
+        if idx < 0 || idx >= self.0.len() as isize {
+            Err(PyIndexError::new_err("array index out of bounds"))
+        } else {
+            Ok(self.0.try_set(idx as usize, value.0)?)
+        }
+    }
+}
+
+#[pyproto]
+impl PySequenceProtocol for PySIArray2 {
+    fn __len__(&self) -> PyResult<usize> {
+        Ok(self.0.len())
+    }
+}
+
+#[pyproto]
+impl PySequenceProtocol for PySIArray3 {
+    fn __len__(&self) -> PyResult<usize> {
+        Ok(self.0.len())
+    }
+}
+
+#[pyproto]
+impl PySequenceProtocol for PySIArray4 {
+    fn __len__(&self) -> PyResult<usize> {
+        Ok(self.0.len())
     }
 }